New INI File
Server Side INI and Root INI will send ini to you
