+++ Scripts/Administration/CurrentExpansion.cs
@@ -10,7 +10,7 @@ namespace Server
 {
 	public class CurrentExpansion
 	{
-		public static readonly Expansion Expansion = Config.GetEnum("Expansion.CurrentExpansion", Expansion.EJ);
+		public static readonly Expansion Expansion = INI.GetEnum("Expansion.CurrentExpansion", Expansion.EJ);
 
 		[CallPriority(Int32.MinValue)]
 		public static void Configure()
