+++ Scripts/Administration/Email.cs
@@ -20,16 +20,16 @@ namespace Server.Misc
         * Example:
         *  public static readonly string SpeechLogPageAddresses = "first@email.here,second@email.here,third@email.here";
         */
-        public static readonly string EmailServer = Config.Get("Email.EmailServer", default(string));
-        public static readonly int EmailPort = Config.Get("Email.EmailPort", 25);
-        public static readonly bool EmailSsl = Config.Get("Email.EmailSsl", false);
+        public static readonly string EmailServer = INI.Get("Email.EmailServer", default(string));
+        public static readonly int EmailPort = INI.Get("Email.EmailPort", 25);
+        public static readonly bool EmailSsl = INI.Get("Email.EmailSsl", false);
 
-        public static readonly string FromAddress = Config.Get("Email.FromAddress", default(string));
-        public static readonly string CrashAddresses = Config.Get("Email.CrashAddresses", default(string));
-        public static readonly string SpeechLogPageAddresses = Config.Get("Email.SpeechLogPageAddresses", default(string));
+        public static readonly string FromAddress = INI.Get("Email.FromAddress", default(string));
+        public static readonly string CrashAddresses = INI.Get("Email.CrashAddresses", default(string));
+        public static readonly string SpeechLogPageAddresses = INI.Get("Email.SpeechLogPageAddresses", default(string));
 
-        public static readonly string EmailUsername = Config.Get("Email.EmailUsername", default(string));
-        public static readonly string EmailPassword = Config.Get("Email.EmailPassword", default(string));
+        public static readonly string EmailUsername = INI.Get("Email.EmailUsername", default(string));
+        public static readonly string EmailPassword = INI.Get("Email.EmailPassword", default(string));
 
         private static readonly Regex _pattern = new Regex(@"^[a-z0-9.+_-]+@([a-z0-9-]+\.)+[a-z]+$", RegexOptions.Compiled | RegexOptions.IgnoreCase);
