+++ Scripts/Administration/World/Siege.cs
@@ -15,8 +15,8 @@ namespace Server
 {
 	public static class Siege
 	{
-		public static bool SiegeShard = Config.Get("Siege.IsSiege", false);
-		public static int CharacterSlots = Config.Get("Siege.CharacterSlots", 1);
+		public static bool SiegeShard = INI.Get("Siege.IsSiege", false);
+		public static int CharacterSlots = INI.Get("Siege.CharacterSlots", 1);
 		public static string FilePath = Path.Combine("Saves", "Siege.bin");
 
 		public static int StatsPerDay = 15;
