+++ Scripts/Expansion/Original UO/Items/Multis/BaseHouse.cs
@@ -20,7 +20,7 @@ namespace Server.Multis
 {
     public abstract class BaseHouse : BaseMulti
     {
-        public static int AccountHouseLimit { get; } = Config.Get("Housing.AccountHouseLimit", 1);
+        public static int AccountHouseLimit { get; } = INI.Get("Housing.AccountHouseLimit", 1);
 
         public static bool NewVendorSystem { get { return Core.AOS; } } // Is new player vendor system enabled?
