+++ Scripts/Expansion/UOR/Services/Factions/Core/Faction.cs
@@ -84,7 +84,7 @@ namespace Server.Factions
 
         public static void Configure()
         {
-            NewCoMLocation = Config.Get("Factions.NewCoMLocation", true);
+            NewCoMLocation = INI.Get("Factions.NewCoMLocation", true);
             Enabled = !Server.Engines.VvV.ViceVsVirtueSystem.Enabled;
 
             EventSink.Login += OnLogin;
