+++ Scripts/Expansion/UOTD/ChampionSystem/ChampionSystem.cs
@@ -71,21 +71,21 @@ namespace Server.Engines.CannedEvil
 
 		public static void Configure()
 		{
-			m_Enabled = Config.Get("Champions.Enabled", true);
-			m_RotateDelay = Config.Get("Champions.RotateDelay", TimeSpan.FromDays(1.0d));
-			m_GoldShowerPiles = Config.Get("Champions.GoldPiles", 50);
-			m_GoldShowerMinAmount = Config.Get("Champions.GoldMin", 4000);
-			m_GoldShowerMaxAmount = Config.Get("Champions.GoldMax", 5500);
-			m_HarrowerGoldPiles = Config.Get("Champions.HarrowerGoldPiles", 75);
-			m_HarrowerGoldMinAmount = Config.Get("Champions.HarrowerGoldMin", 5000);
-			m_HarrowerGoldMaxAmount = Config.Get("Champions.HarrowerGoldMax", 10000);
-			m_PowerScrollAmount = Config.Get("Champions.PowerScrolls", 6);
-			m_StatScrollAmount = Config.Get("Champions.StatScrolls", 16);
-			m_ScrollChance = Config.Get("Champions.ScrollChance", 0.1d) / 100.0d;
-			m_TranscendenceChance = Config.Get("Champions.TranscendenceChance", 50.0d) / 100.0d;
-			int rank2 = Config.Get("Champions.Rank2RedSkulls", 5);
-			int rank3 = Config.Get("Champions.Rank3RedSkulls", 10);
-			int rank4 = Config.Get("Champions.Rank4RedSkulls", 10);
+			m_Enabled = INI.Get("Champions.Enabled", true);
+			m_RotateDelay = INI.Get("Champions.RotateDelay", TimeSpan.FromDays(1.0d));
+			m_GoldShowerPiles = INI.Get("Champions.GoldPiles", 50);
+			m_GoldShowerMinAmount = INI.Get("Champions.GoldMin", 4000);
+			m_GoldShowerMaxAmount = INI.Get("Champions.GoldMax", 5500);
+			m_HarrowerGoldPiles = INI.Get("Champions.HarrowerGoldPiles", 75);
+			m_HarrowerGoldMinAmount = INI.Get("Champions.HarrowerGoldMin", 5000);
+			m_HarrowerGoldMaxAmount = INI.Get("Champions.HarrowerGoldMax", 10000);
+			m_PowerScrollAmount = INI.Get("Champions.PowerScrolls", 6);
+			m_StatScrollAmount = INI.Get("Champions.StatScrolls", 16);
+			m_ScrollChance = INI.Get("Champions.ScrollChance", 0.1d) / 100.0d;
+			m_TranscendenceChance = INI.Get("Champions.TranscendenceChance", 50.0d) / 100.0d;
+			int rank2 = INI.Get("Champions.Rank2RedSkulls", 5);
+			int rank3 = INI.Get("Champions.Rank3RedSkulls", 10);
+			int rank4 = INI.Get("Champions.Rank4RedSkulls", 10);
 			for (int i = 0; i < m_Rank.Length; ++i)
 			{
 				if (i < rank2)
@@ -97,10 +97,10 @@ namespace Server.Engines.CannedEvil
 				else
 					m_Rank[i] = 3;
 			}
-			m_MaxKill[0] = Config.Get("Champions.Rank1MaxKills", 256);
-			m_MaxKill[1] = Config.Get("Champions.Rank2MaxKills", 128);
-			m_MaxKill[2] = Config.Get("Champions.Rank3MaxKills", 64);
-			m_MaxKill[3] = Config.Get("Champions.Rank4MaxKills", 32);
+			m_MaxKill[0] = INI.Get("Champions.Rank1MaxKills", 256);
+			m_MaxKill[1] = INI.Get("Champions.Rank2MaxKills", 128);
+			m_MaxKill[2] = INI.Get("Champions.Rank3MaxKills", 64);
+			m_MaxKill[3] = INI.Get("Champions.Rank4MaxKills", 32);
 			EventSink.WorldLoad += EventSink_WorldLoad;
 			EventSink.WorldSave += EventSink_WorldSave;
 		}
