+++ Scripts/Mechanics/XmlSpawner/XmlSpawner Core/XmlSpawner2.cs
@@ -51,9 +51,9 @@ namespace Server.Mobiles
 
 		public static void Configure()
 		{
-			m_XmlPoints = Config.Get("XmlSpawner2.Points", false);
-			m_XmlFactions = Config.Get("XmlSpawner2.Factions", false);
-			m_XmlSockets = Config.Get("XmlSpawner2.Sockets", false);
+			m_XmlPoints = INI.Get("XmlSpawner2.Points", false);
+			m_XmlFactions = INI.Get("XmlSpawner2.Factions", false);
+			m_XmlSockets = INI.Get("XmlSpawner2.Sockets", false);
 		}
 
 		#region Type declarations
